+++ src/CleanRegex/Builder/PatternBuilder.php
@@ -1,10 +1,10 @@
 <?php
 namespace TRegx\CleanRegex\Builder;
 
-use TRegx\CleanRegex\Internal\Delimiter\Strategy\StandardStrategy;
-use TRegx\CleanRegex\Internal\Prepared\InjectParser;
-use TRegx\CleanRegex\Internal\Prepared\MaskParser;
-use TRegx\CleanRegex\Internal\Prepared\PrepareFacade;
+use TRegx\CleanRegex\Internal\Prepared\Expression\Mask;
+use TRegx\CleanRegex\Internal\Prepared\Expression\Template;
+use TRegx\CleanRegex\Internal\Prepared\Figure\InjectFigures;
+use TRegx\CleanRegex\Internal\Prepared\Orthography\StandardOrthography;
 use TRegx\CleanRegex\Pattern;
 
 class PatternBuilder
@@ -16,16 +16,18 @@ class PatternBuilder
 
     public function inject(string $input, array $values, string $flags = null): Pattern
     {
-        return PrepareFacade::build(new InjectParser($input, $values), new StandardStrategy($flags ?? ''));
+        $template = new Template(new StandardOrthography($input, $flags ?? ''), new InjectFigures($values));
+        return new Pattern($template->definition());
     }
 
     public function mask(string $mask, array $keywords, string $flags = null): Pattern
     {
-        return PrepareFacade::build(new MaskParser($mask, $keywords), new StandardStrategy($flags ?? ''));
+        $mask = new Mask($mask, $keywords, $flags ?? '');
+        return new Pattern($mask->definition());
     }
 
     public function template(string $pattern, string $flags = null): TemplateBuilder
     {
-        return new TemplateBuilder($pattern, new StandardStrategy($flags ?? ''), []);
+        return new TemplateBuilder(new StandardOrthography($pattern, $flags ?? ''), []);
     }
 }