+++ src/CleanRegex/Builder/PcrePatternBuilder.php
@@ -1,20 +1,21 @@
 <?php
 namespace TRegx\CleanRegex\Builder;
 
-use TRegx\CleanRegex\Internal\Delimiter\Strategy\PcreStrategy;
-use TRegx\CleanRegex\Internal\Prepared\InjectParser;
-use TRegx\CleanRegex\Internal\Prepared\PrepareFacade;
+use TRegx\CleanRegex\Internal\Prepared\Expression\Template;
+use TRegx\CleanRegex\Internal\Prepared\Figure\InjectFigures;
+use TRegx\CleanRegex\Internal\Prepared\Orthography\PcreOrthography;
 use TRegx\CleanRegex\Pattern;
 
 class PcrePatternBuilder
 {
     public function inject(string $input, array $values): Pattern
     {
-        return PrepareFacade::build(new InjectParser($input, $values), new PcreStrategy());
+        $build = new Template(new PcreOrthography($input), new InjectFigures($values));
+        return new Pattern($build->definition());
     }
 
     public function template(string $pattern): TemplateBuilder
     {
-        return new TemplateBuilder($pattern, new PcreStrategy(), []);
+        return new TemplateBuilder(new PcreOrthography($pattern), []);
     }
 }