+++ src/CleanRegex/Builder/TemplateBuilder.php
@@ -1,28 +1,24 @@
 <?php
 namespace TRegx\CleanRegex\Builder;
 
-use TRegx\CleanRegex\Exception\TemplateFormatException;
-use TRegx\CleanRegex\Internal\Delimiter\Strategy\DelimiterStrategy;
-use TRegx\CleanRegex\Internal\Prepared\PrepareFacade;
+use TRegx\CleanRegex\Internal\Prepared\Expression\Template;
+use TRegx\CleanRegex\Internal\Prepared\Figure\TokenFigures;
+use TRegx\CleanRegex\Internal\Prepared\Orthography\Orthography;
 use TRegx\CleanRegex\Internal\Prepared\Template\LiteralToken;
 use TRegx\CleanRegex\Internal\Prepared\Template\MaskToken;
 use TRegx\CleanRegex\Internal\Prepared\Template\Token;
-use TRegx\CleanRegex\Internal\Prepared\TemplateParser;
 use TRegx\CleanRegex\Pattern;
 
 class TemplateBuilder
 {
-    /** @var string */
-    private $pattern;
-    /** @var DelimiterStrategy */
-    private $strategy;
+    /** @var Orthography */
+    private $orthography;
     /** @var Token[] */
     private $tokens;
 
-    public function __construct(string $pattern, DelimiterStrategy $strategy, array $tokens)
+    public function __construct(Orthography $orthography, array $tokens)
     {
-        $this->pattern = $pattern;
-        $this->strategy = $strategy;
+        $this->orthography = $orthography;
         $this->tokens = $tokens;
     }
 
@@ -38,24 +34,12 @@ class TemplateBuilder
 
     private function next(Token $token): TemplateBuilder
     {
-        return new TemplateBuilder($this->pattern, $this->strategy, \array_merge($this->tokens, [$token]));
+        return new TemplateBuilder($this->orthography, \array_merge($this->tokens, [$token]));
     }
 
     public function build(): Pattern
     {
-        $this->validateTokensAndMethods();
-        return PrepareFacade::build(new TemplateParser($this->pattern, $this->tokens), $this->strategy);
-    }
-
-    private function validateTokensAndMethods(): void
-    {
-        $placeholders = \preg_match_all('/@/', $this->pattern);
-        $tokens = \count($this->tokens);
-        if ($placeholders < $tokens) {
-            throw TemplateFormatException::insufficient($placeholders, $tokens);
-        }
-        if ($placeholders > $tokens) {
-            throw TemplateFormatException::superfluous($placeholders, $tokens);
-        }
+        $build = new Template($this->orthography, new TokenFigures($this->tokens));
+        return new Pattern($build->definition());
     }
 }