+++ test/Feature/CleanRegex/_prepared/template/PatternTest.php
@@ -1,0 +1,103 @@
+<?php
+namespace Test\Feature\TRegx\CleanRegex\_prepared\template;
+
+use PHPUnit\Framework\TestCase;
+use Test\Utils\AssertsPattern;
+use TRegx\CleanRegex\Internal\Prepared\Figure\PlaceholderFigureException;
+use TRegx\CleanRegex\Pattern;
+
+class PatternTest extends TestCase
+{
+    use AssertsPattern;
+
+    /**
+     * @test
+     * @dataProvider templatesWithPlaceholder
+     * @param string $pattern
+     * @param string $expected
+     */
+    public function shouldUsePlaceholder(string $pattern, string $expected)
+    {
+        // when
+        $pattern = Pattern::template($pattern)->literal('X')->build();
+
+        // then
+        $this->assertSamePattern($expected, $pattern);
+    }
+
+    public function templatesWithPlaceholder(): array
+    {
+        return [
+            'standard'                               => ['You/her @ her?', '#You/her X her?#'],
+            'comment (but no "x" flag)'              => ["You/her #@\n her?", "%You/her #X\n her?%"],
+            'comment ("x" flag, but also "-x" flag)' => ["You/her (?x:(?-x:#@\n)) her?", "%You/her (?x:(?-x:#X\n)) her?%"],
+        ];
+    }
+
+    /**
+     * @test
+     * @dataProvider templatesWithoutPlaceholders
+     * @param string $pattern
+     * @param string $expected
+     */
+    public function shouldNotMistakeLiteralForPlaceholder(string $pattern, string $expected)
+    {
+        // when
+        $pattern = Pattern::template($pattern)->build();
+
+        // then
+        $this->assertSamePattern($expected, $pattern);
+    }
+
+    public function templatesWithoutPlaceholders(): array
+    {
+        return [
+            "placeholder '@' in []"      => ['You/her [@] her?', '#You/her [@] her?#'],
+            "placeholder '@' in \Q\E"    => ['You/her \Q@\E her?', '#You/her \Q@\E her?#'],
+            "placeholder '@' escaped"    => ['You/her \@ her?', '#You/her \@ her?#'],
+            "placeholder '@' in comment" => ["You/her (?x:#@\n) her?", "%You/her (?x:#@\n) her?%"],
+            "placeholder '@' in control" => ["You/her \c@ her?", "#You/her \c@ her?#"],
+        ];
+    }
+
+    /**
+     * @test
+     */
+    public function shouldNotMistakePlaceholderInCommentInExtendedMode()
+    {
+        // when
+        $pattern = Pattern::template("You/her #@\n her?", 'x')->build();
+
+        // then
+        $this->assertSamePattern("%You/her #@\n her?%x", $pattern);
+    }
+
+    /**
+     * @test
+     */
+    public function shouldUsePlaceholderInCommentInExtendedMode_butExtendedModeIsSwitchedOff()
+    {
+        // when
+        $pattern = Pattern::template("You/her (?-x:#@\n) her?", 'x')->literal('X')->build();
+
+        // then
+        $this->assertSamePattern("%You/her (?-x:#X\n) her?%x", $pattern);
+    }
+
+    /**
+     * @test
+     */
+    public function shouldThrowForSuperfluousTemplatePlaceholder()
+    {
+        // then
+        $this->expectException(PlaceholderFigureException::class);
+        $this->expectExceptionMessage("Found a superfluous figure: string ('bar'). Used 1 placeholders, but 3 figures supplied.");
+
+        // when
+        Pattern::builder()->template('You/her, (are|is) @ (you|her)')
+            ->literal('foo')
+            ->literal('bar')
+            ->literal('cat')
+            ->build();
+    }
+}